backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
langchain==0.3.7
langchain-groq==0.2.1
langgraph==0.2.39
langchain-core==0.3.15
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
alembic==1.13.1
python-multipart==0.0.9